refactor(backend): split document course plan builders
