+++ docs/architecture/WIII_CODEBASE_MAP.md
@@ -61,7 +61,7 @@ The normal chat turn should be read in this order:
 | Streaming behavior | `maritime-ai-service/app/api/v1/chat_stream.py`, then `maritime-ai-service/app/services/chat_stream_coordinator.py` |
 | Direct response node | `maritime-ai-service/app/engine/multi_agent/direct_node_runtime.py` (main direct node)<br>Deterministic fast responses: `direct_node_fast_response_runtime.py`<br>Uploaded-document preview preflight: `direct_node_document_preview_runtime.py`<br>Deterministic thinking snapshots: `direct_node_thinking_snapshot.py`<br>Fast-path helper modules: `direct_node_*_runtime.py` |
 | Direct answer streaming | `maritime-ai-service/app/engine/multi_agent/direct_execution.py` (provider invocation and fallback shell)<br>Visible stream text/thinking helpers: `direct_stream_text_runtime.py`<br>Opening/wait surfaces: `direct_opening_runtime.py`, `direct_wait_surface.py` |
-| Direct tool loop | `maritime-ai-service/app/engine/multi_agent/direct_tool_rounds_runtime.py` (main orchestration)<br>Runtime/provider bindings: `direct_runtime_bindings.py`<br>Generic dispatch: `direct_tool_dispatch_runtime.py`<br>Final synthesis helpers: `direct_final_synthesis_runtime.py`<br>Pointy policy: `direct_pointy_runtime.py`<br>Visual turn policy: `direct_visual_tool_policy_runtime.py`<br>Web-search policy: `direct_web_search_policy.py`<br>Document host-action shortcuts: `direct_document_host_action_runtime.py`<br>Uploaded-document preview/course payloads: `direct_document_preview_payloads.py`<br>Uploaded-document text shaping: `direct_document_preview_text.py`<br>Uploaded-document source refs: `direct_document_source_refs.py`<br>Uploaded-document domain course plans: `direct_document_course_domain_plans.py`<br>Message builders: `direct_tool_message_runtime.py`<br>Session-memory fast paths: `direct_session_memory_runtime.py` |
+| Direct tool loop | `maritime-ai-service/app/engine/multi_agent/direct_tool_rounds_runtime.py` (main orchestration)<br>Runtime/provider bindings: `direct_runtime_bindings.py`<br>Generic dispatch: `direct_tool_dispatch_runtime.py`<br>Final synthesis helpers: `direct_final_synthesis_runtime.py`<br>Pointy policy: `direct_pointy_runtime.py`<br>Visual turn policy: `direct_visual_tool_policy_runtime.py`<br>Web-search policy: `direct_web_search_policy.py`<br>Document host-action shortcuts: `direct_document_host_action_runtime.py`<br>Uploaded-document preview/course payloads: `direct_document_preview_payloads.py`<br>Uploaded-document text shaping: `direct_document_preview_text.py`<br>Uploaded-document source refs: `direct_document_source_refs.py`<br>Uploaded-document course plans: `direct_document_course_lms_plan.py`, `direct_document_course_maritime_plans.py` (`direct_document_course_domain_plans.py` is compatibility exports)<br>Message builders: `direct_tool_message_runtime.py`<br>Session-memory fast paths: `direct_session_memory_runtime.py` |
 | Native stream/provider quirks | `maritime-ai-service/app/engine/multi_agent/openai_stream_runtime.py` |
 | Routing and supervisor behavior | `maritime-ai-service/app/engine/multi_agent/supervisor*.py` |
 | Direct prompt assembly | `maritime-ai-service/app/engine/multi_agent/direct_prompts.py` (main assembly)<br>Tool binding: `direct_prompt_tool_binding.py`<br>Tool-context prompt builders: `direct_prompt_tool_context.py`<br>Turn contract and forced-skill prompt helpers: `direct_prompt_turn_contracts.py` |
