+++ docs/operations/WIII_RUNTIME_CLEANUP_AUDIT_2026-05-19.md
@@ -6,7 +6,7 @@ Owner: Project leadership
 
 Issue: #411
 
-Follow-up issues: #413, #415, #417, #419, #421, #423, #425, #427, #429, #431, #433, #435, #437, #439, #441, #477, #479, #481, #483, #485, #487, #489, #491, #493, #495, #497 (owner: Architecture Maintainers)
+Follow-up issues: #413, #415, #417, #419, #421, #423, #425, #427, #429, #431, #433, #435, #437, #439, #441, #477, #479, #481, #483, #485, #487, #489, #491, #493, #495, #497, #499 (owner: Architecture Maintainers)
 
 ## Purpose
 
@@ -133,6 +133,11 @@ without broad deletes, secret exposure, or unreviewed product behavior changes.
   native chunk part extraction, duplicate answer-delta comparison, visible
   thinking cleanup/alignment, and pseudo-stream answer chunking outside the
   provider fallback shell.
+- Follow-up #499 split deterministic uploaded-document course plan builders by
+  domain into `direct_document_course_lms_plan.py` and
+  `direct_document_course_maritime_plans.py`, leaving
+  `direct_document_course_domain_plans.py` as a compatibility export surface for
+  existing document preview/tool-loop imports.
 
 ## Preserved Intentionally
 
@@ -174,8 +179,8 @@ backups, data PDFs, or local skill folders.
 - `direct_tool_rounds_runtime.py` is now a smaller orchestration shell. Pointy,
   explicit web-search policy, deterministic document host-action execution,
   uploaded-document preview/course payload shell, uploaded-document text
-  shaping, document source refs, domain course plan builders, visual-turn
-  policy, graph/runtime binding resolution, per-turn provider policy, message
+  shaping, document source refs, domain-specific course plan builder modules,
+  visual-turn policy, graph/runtime binding resolution, per-turn provider policy, message
   builders, generic tool dispatch, final synthesis helper construction, final
   synthesis execution, post-tool convergence policy, follow-up LLM
   selection/invocation, response finalization, post-tool search-template
@@ -324,3 +329,8 @@ In follow-up #497, the direct execution streaming regression set passed with 22
 tests after moving visible stream text helpers into
 `direct_stream_text_runtime.py`. Targeted ruff checks passed for the direct
 execution compatibility wrapper and new helper module.
+In follow-up #499, the document preview/tool-loop regression set passed with
+93 tests after splitting uploaded-document course plan builders into LMS and
+maritime modules. The focused host UI/tool surface regression set passed with
+14 tests. Targeted ruff checks, repository `ruff check app/ --select=E9,F63,F7`,
+and `git diff --check` also passed for the domain plan builder split.
